# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Fortnite-api",
    "description": "REST Fortnite API. Get stats, score, kills...",
    "version": "1.0.0",
    "contact": {
      "name": "SkYNewZ",
      "url": "https://www.lemairepro.fr",
      "email": "quentin@lemairepro.fr"
    },
    "license": {
      "name": "GNU General Public License v3.0",
      "url": "https://en.wikipedia.org/wiki/GNU_General_Public_License"
    }
  },
  "host": "HOSTNAME",
  "basePath": "/api",
  "schemes": [
    "https",
    "http"
  ],
...
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/lookup/{username}": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "summary": "Return user info",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/lookup": {
      "get": {
        "summary": "Return array of multiple user info",
        "tags": [
          "Lookup"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "Comma-separated list of users's id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "Unique id in EpicGames server"
        },
        "displayName": {
          "type": "string",
          "description": "In-game display name"
        }
      }
    },
...
  }
}